move more functionality to UserPoolEntryUpdater
