+++ cognito-pre-token-generation/src/main/java/no/unit/nva/cognito/TriggerHandler.java
@@ -1,10 +1,25 @@
 package no.unit.nva.cognito;
 
-import com.amazonaws.services.cognitoidp.model.AttributeType;
+import static java.util.Objects.nonNull;
+import static no.unit.nva.cognito.Constants.DYNAMODB_CLIENT;
+import static no.unit.nva.cognito.Constants.ENVIRONMENT;
+import static no.unit.nva.cognito.Constants.ID_NAMESPACE_VALUE;
+import static no.unit.nva.cognito.util.OrgNumberCleaner.removeCountryPrefix;
+import static no.unit.nva.customer.ObjectMapperConfig.objectMapper;
+import static nva.commons.core.StringUtils.isNotBlank;
+import static nva.commons.core.attempt.Try.attempt;
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
+import java.net.URLDecoder;
+import java.nio.charset.StandardCharsets;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.stream.Collectors;
 import no.unit.nva.cognito.model.CustomerResponse;
 import no.unit.nva.cognito.model.Event;
+import no.unit.nva.cognito.model.Request;
 import no.unit.nva.cognito.model.UserAttributes;
 import no.unit.nva.cognito.service.CustomerApi;
 import no.unit.nva.cognito.service.CustomerDbClient;
@@ -15,7 +30,6 @@ import no.unit.nva.cognito.service.UserService;
 import no.unit.nva.customer.model.CustomerMapper;
 import no.unit.nva.customer.service.impl.DynamoDBCustomerService;
 import no.unit.nva.database.DatabaseServiceImpl;
-import no.unit.nva.useraccessmanagement.model.RoleDto;
 import no.unit.nva.useraccessmanagement.model.UserDto;
 import nva.commons.core.JacocoGenerated;
 import nva.commons.core.JsonUtils;
@@ -24,38 +38,10 @@ import nva.commons.core.attempt.Try;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.net.URLDecoder;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
-import static java.util.Objects.nonNull;
-import static no.unit.nva.cognito.Constants.DYNAMODB_CLIENT;
-import static no.unit.nva.cognito.Constants.ENVIRONMENT;
-import static no.unit.nva.cognito.Constants.ID_NAMESPACE_VALUE;
-import static no.unit.nva.cognito.util.OrgNumberCleaner.removeCountryPrefix;
-import static no.unit.nva.customer.ObjectMapperConfig.objectMapper;
-import static nva.commons.core.StringUtils.isNotBlank;
-import static nva.commons.core.attempt.Try.attempt;
-
 public class TriggerHandler implements RequestHandler<Map<String, Object>, Map<String, Object>> {
 
-    public static final String CUSTOM_APPLICATION_ROLES = "custom:applicationRoles";
-    public static final String CUSTOM_APPLICATION = "custom:application";
-    public static final String CUSTOM_CUSTOMER_ID = "custom:customerId";
-    public static final String CUSTOM_IDENTIFIERS = "custom:identifiers";
-    public static final String CUSTOM_CRISTIN_ID = "custom:cristinId";
-    public static final String CUSTOM_APPLICATION_ACCESS_RIGHTS = "custom:accessRights";
-
     public static final String COMMA_DELIMITER = ",";
-    public static final String FEIDE_PREFIX = "feide:";
-    public static final String NVA = "NVA";
+
     public static final String BIBSYS_HOST = "@bibsys.no";
     public static final String EMPTY_STRING = "";
     public static final int START_OF_STRING = 0;
@@ -63,11 +49,10 @@ public class TriggerHandler implements RequestHandler<Map<String, Object>, Map<S
     public static final char AFFILIATION_PART_SEPARATOR = '@';
     public static final String COMMA_SPACE = ", ";
     public static final String COMMA = ",";
-    public static final String APPLICATION_ROLES_MESSAGE = "applicationRoles: ";
     public static final String HOSTED_AFFILIATION_MESSAGE =
-            "Overriding orgNumber({}) with hostedOrgNumber({}) and hostedAffiliation";
+        "Overriding orgNumber({}) with hostedOrgNumber({}) and hostedAffiliation";
     public static final String PROBLEM_DECODING_HOSTED_USERS_AFFILIATION =
-            "Problem decoding hosted users affiliation, using original";
+        "Problem decoding hosted users affiliation, using original";
     private static final Logger logger = LoggerFactory.getLogger(TriggerHandler.class);
     private final UserService userService;
     private final CustomerApi customerApi;
@@ -85,6 +70,21 @@ public class TriggerHandler implements RequestHandler<Map<String, Object>, Map<S
         this.userPoolEntryUpdater = userPoolEntryUpdater;
     }
 
+    @Override
+    public Map<String, Object> handleRequest(Map<String, Object> input, Context context) {
+        long start = System.currentTimeMillis();
+        Event event = parseEventFromInput(input);
+
+
+        UserDetails userDetails = extractUserDetails(event);
+        UserDto user = getAndUpdateUserDetails(userDetails);
+
+        updateUserDetailsInUserPool(userDetails, user);
+
+        logger.info("handleRequest took {} ms", System.currentTimeMillis() - start);
+        return input;
+    }
+
     @JacocoGenerated
     private static CustomerDbClient defaultCustomerDbClient() {
         return new CustomerDbClient(defaultCustomerService(), defaultCustomerMapper());
@@ -93,9 +93,9 @@ public class TriggerHandler implements RequestHandler<Map<String, Object>, Map<S
     @JacocoGenerated
     private static DynamoDBCustomerService defaultCustomerService() {
         return new DynamoDBCustomerService(
-                DYNAMODB_CLIENT,
-                objectMapper,
-                ENVIRONMENT);
+            DYNAMODB_CLIENT,
+            objectMapper,
+            ENVIRONMENT);
     }
 
     @JacocoGenerated
@@ -113,28 +113,12 @@ public class TriggerHandler implements RequestHandler<Map<String, Object>, Map<S
         return new UserDbClient(new DatabaseServiceImpl());
     }
 
-    @Override
-    public Map<String, Object> handleRequest(Map<String, Object> input, Context context) {
-        long start = System.currentTimeMillis();
-        Event event = parseEventFromInput(input);
-
-        String userPoolId = event.getUserPoolId();
-        String userName = event.getUserName();
-        UserDetails userDetails = extractUserDetails(event);
-        UserDto user = getAndUpdateUserDetails(userDetails);
-
-        updateUserDetailsInUserPool(userPoolId, userName, userDetails, user);
-
-        logger.info("handleRequest took {} ms", System.currentTimeMillis() - start);
-        return input;
-    }
-
     private UserDetails extractUserDetails(Event event) {
         UserAttributes userAttributes = event.getRequest().getUserAttributes();
         if (userIsBibsysHosted(userAttributes)) {
             injectInformationForBibsysHostedCustomer(userAttributes);
         }
-        return createUserDetails(userAttributes);
+        return createUserDetails(event);
     }
 
     private void injectInformationForBibsysHostedCustomer(UserAttributes userAttributes) {
@@ -143,11 +127,14 @@ public class TriggerHandler implements RequestHandler<Map<String, Object>, Map<S
         userAttributes.setAffiliation(extractAffiliationFromHostedUSer(userAttributes.getHostedAffiliation()));
     }
 
-    private UserDetails createUserDetails(UserAttributes userAttributes) {
-        return Optional.ofNullable(userAttributes.getOrgNumber())
-                .flatMap(orgNum -> mapOrgNumberToCustomer(removeCountryPrefix(orgNum)))
-                .map(customer -> new UserDetails(userAttributes, customer))
-                .orElse(new UserDetails(userAttributes));
+    private UserDetails createUserDetails(Event event) {
+        return Optional.of(event)
+            .map(Event::getRequest)
+            .map(Request::getUserAttributes)
+            .map(UserAttributes::getOrgNumber)
+            .flatMap(orgNum -> mapOrgNumberToCustomer(removeCountryPrefix(orgNum)))
+            .map(customer -> new UserDetails(event, customer))
+            .orElse(new UserDetails(event));
     }
 
     /**
@@ -161,88 +148,35 @@ public class TriggerHandler implements RequestHandler<Map<String, Object>, Map<S
         return JsonUtils.objectMapper.convertValue(input, Event.class);
     }
 
-    private void updateUserDetailsInUserPool(String userPoolId,
-                                             String userName,
-                                             UserDetails userDetails,
-                                             UserDto user) {
+    private void updateUserDetailsInUserPool(UserDetails userDetails,UserDto user) {
         long start = System.currentTimeMillis();
-        List<AttributeType> cognitoUserAttributes = createUserAttributes(userDetails, user);
-        userPoolEntryUpdater.updateUserAttributes(userPoolId, userName, cognitoUserAttributes);
+        userPoolEntryUpdater.updateUserAttributes(userDetails,user);
         logger.info("updateUserDetailsInUserPool took {} ms", System.currentTimeMillis() - start);
     }
 
     private UserDto getAndUpdateUserDetails(UserDetails userDetails) {
         return userService.getUser(userDetails.getFeideId())
-                .map(attempt(user -> userService.updateUser(user, userDetails)))
-                .map(Try::orElseThrow)
-                .orElseGet(() -> userService.createUser(userDetails));
+            .map(attempt(user -> userService.updateUser(user, userDetails)))
+            .map(Try::orElseThrow)
+            .orElseGet(() -> userService.createUser(userDetails));
     }
 
     private Optional<CustomerResponse> mapOrgNumberToCustomer(String orgNumber) {
         return customerApi.getCustomer(orgNumber);
     }
 
-    private List<AttributeType> createUserAttributes(UserDetails userDetails, UserDto user) {
-        List<AttributeType> userAttributeTypes = new ArrayList<>();
-
-        if (user.getInstitution() != null) {
-            userAttributeTypes.add(toAttributeType(CUSTOM_CUSTOMER_ID, user.getInstitution()));
-        }
-        userDetails.getCristinId()
-                .ifPresent(cristinId -> userAttributeTypes.add(toAttributeType(CUSTOM_CRISTIN_ID, cristinId)));
-
-        userAttributeTypes.add(toAttributeType(CUSTOM_APPLICATION, NVA));
-        userAttributeTypes.add(toAttributeType(CUSTOM_IDENTIFIERS, FEIDE_PREFIX + userDetails.getFeideId()));
-
-        String applicationRoles = applicationRolesString(user);
-        userAttributeTypes.add(toAttributeType(CUSTOM_APPLICATION_ROLES, applicationRoles));
-
-        String accessRightsString = accessRightsString(user);
-        userAttributeTypes.add(toAttributeType(CUSTOM_APPLICATION_ACCESS_RIGHTS, accessRightsString));
-
-        return userAttributeTypes;
-    }
-
-    private String accessRightsString(UserDto user) {
-        if (!user.getAccessRights().isEmpty()) {
-            return toCsv(user.getAccessRights(), s -> s);
-        } else {
-            return EMPTY_STRING;
-        }
-    }
-
-    private String applicationRolesString(UserDto user) {
-        String applicationRoles = toCsv(user.getRoles(), RoleDto::getRoleName);
-        logger.info(APPLICATION_ROLES_MESSAGE + applicationRoles);
-        return applicationRoles;
-    }
-
-    private AttributeType toAttributeType(String name, String value) {
-        AttributeType attributeType = new AttributeType();
-        attributeType.setName(name);
-        attributeType.setValue(value);
-        return attributeType;
-    }
-
-    private <T> String toCsv(Collection<T> roles, Function<T, String> stringRepresentation) {
-        return roles
-                .stream()
-                .map(stringRepresentation)
-                .collect(Collectors.joining(COMMA_DELIMITER));
-    }
-
     private boolean userIsBibsysHosted(UserAttributes userAttributes) {
         return userAttributes.getFeideId().endsWith(BIBSYS_HOST)
-                && nonNull(userAttributes.getHostedOrgNumber());
+               && nonNull(userAttributes.getHostedOrgNumber());
     }
 
     private String extractAffiliationFromHostedUSer(String hostedAffiliation) {
 
         List<String> shortenedAffiliations = Arrays.stream(hostedAffiliation.split(COMMA))
-                .map(this::decodeAffiliation)
-                .map(this::extractAffiliation)
-                .map(String::strip)
-                .collect(Collectors.toList());
+            .map(this::decodeAffiliation)
+            .map(this::extractAffiliation)
+            .map(String::strip)
+            .collect(Collectors.toList());
 
         return String.join(COMMA_SPACE, shortenedAffiliations).concat(TRAILING_BRACKET);
     }
@@ -257,7 +191,7 @@ public class TriggerHandler implements RequestHandler<Map<String, Object>, Map<S
 
     private String decodeAffiliation(String hostedAffiliation) {
         return attempt(() -> URLDecoder.decode(hostedAffiliation, StandardCharsets.UTF_8))
-                .orElse(fail -> logWarningAndReturnOriginalValue(fail, hostedAffiliation));
+            .orElse(fail -> logWarningAndReturnOriginalValue(fail, hostedAffiliation));
     }
 
     private String logWarningAndReturnOriginalValue(Failure<String> fail, String hostedAffiliation) {
