+++ cognito-pre-token-generation/src/main/java/no/unit/nva/cognito/service/UserDetails.java
@@ -2,6 +2,7 @@ package no.unit.nva.cognito.service;
 
 import java.util.Optional;
 import no.unit.nva.cognito.model.CustomerResponse;
+import no.unit.nva.cognito.model.Event;
 import no.unit.nva.cognito.model.UserAttributes;
 import nva.commons.core.StringUtils;
 
@@ -14,23 +15,39 @@ public class UserDetails {
     private final String customerId;
     private final String affiliation;
     private final String cristinId;
+    private final String cognitoUserName;
+    private final String cognitoUserPool;
 
-    public UserDetails(UserAttributes userAttributes, CustomerResponse customer) {
+    public UserDetails(Event event, CustomerResponse customer) {
+        UserAttributes userAttributes = event.getRequest().getUserAttributes();
         this.feideId = userAttributes.getFeideId();
         this.affiliation = nonBlankOrNull(userAttributes.getAffiliation());
         this.customerId = extractCustomerId(customer);
         this.cristinId = extractCristinId(customer);
         this.givenName = userAttributes.getGivenName();
         this.familyName = userAttributes.getFamilyName();
+        this.cognitoUserName = event.getUserName();
+        this.cognitoUserPool = event.getUserPoolId();
     }
 
-    public UserDetails(UserAttributes userAttributes) {
+    public UserDetails(Event event) {
+        UserAttributes userAttributes = event.getRequest().getUserAttributes();
         this.feideId = userAttributes.getFeideId();
         this.affiliation = nonBlankOrNull(userAttributes.getAffiliation());
         this.customerId = NO_CUSTOMER_INFO;
         this.cristinId = NO_CUSTOMER_INFO;
         this.givenName = userAttributes.getGivenName();
         this.familyName = userAttributes.getFamilyName();
+        this.cognitoUserPool = event.getUserPoolId();
+        this.cognitoUserName = event.getUserName();
+    }
+
+    public String getCognitoUserName() {
+        return cognitoUserName;
+    }
+
+    public String getCognitoUserPool() {
+        return cognitoUserPool;
     }
 
     public Optional<String> getCristinId() {
