+++ customer-testing/src/main/java/no/unit/nva/customer/testing/TestHeaders.java
@@ -1,13 +1,13 @@
 package no.unit.nva.customer.testing;
 
+import nva.commons.apigateway.ContentTypes;
 import nva.commons.core.JacocoGenerated;
 
 import java.util.Map;
 
 import static nva.commons.apigateway.ApiGatewayHandler.ACCESS_CONTROL_ALLOW_ORIGIN;
 import static nva.commons.apigateway.ApiGatewayHandler.CONTENT_TYPE;
-import static org.apache.http.HttpHeaders.ACCEPT;
-import static org.apache.http.entity.ContentType.APPLICATION_JSON;
+import static nva.commons.apigateway.HttpHeaders.ACCEPT;
 
 @JacocoGenerated
 public class TestHeaders {
@@ -21,8 +21,8 @@ public class TestHeaders {
      */
     public static Map<String,String> getRequestHeaders() {
         return Map.of(
-                CONTENT_TYPE, APPLICATION_JSON.getMimeType(),
-                ACCEPT, APPLICATION_JSON.getMimeType());
+            CONTENT_TYPE, ContentTypes.APPLICATION_JSON,
+            ACCEPT, ContentTypes.APPLICATION_JSON);
     }
 
     /**
@@ -31,7 +31,7 @@ public class TestHeaders {
      */
     public static  Map<String,String> getResponseHeaders() {
         return Map.of(
-                CONTENT_TYPE, APPLICATION_JSON.getMimeType(),
+                CONTENT_TYPE,  ContentTypes.APPLICATION_JSON,
                 ACCESS_CONTROL_ALLOW_ORIGIN, WILDCARD
         );
     }
